UI Added
UI and score now implemented, the game is nearly playable.
set up and ready for the vehicle creator

+++ 3D/3D game (called Testing_Tool)/Testing_Tool/FPMode.pde
@@ -1,7 +1,7 @@
 void thirdPersonMode()
 {
-    int px = CarArray[0].getX();
-    int py = CarArray[0].getY();
+    float px = CarArray[0].getX();
+    float py = CarArray[0].getY();
     float pt = CarArray[0].getT();
   
     //Dimensions of Car
